refactors to use slices
