+++ internal/cmd/versions_test.go
@@ -12,165 +12,184 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package cmd_test
+package cmd
 
 import (
-	"fmt"
 	"os"
 	"path/filepath"
 	"testing"
-	"time"
 
 	"github.com/stretchr/testify/require"
 
+	"github.com/tetratelabs/getenvoy/internal/test/morerequire"
 	"github.com/tetratelabs/getenvoy/internal/version"
 )
 
-func TestGetEnvoyVersions_NothingYet(t *testing.T) {
-	o, cleanup := setupTest(t)
-	defer cleanup()
+func TestGetInstalledVersions(t *testing.T) {
+	homeDir, removeHomeDir := morerequire.RequireNewTempDir(t)
+	defer removeHomeDir()
 
-	c, stdout, stderr := newApp(o)
-	o.Out = stdout
-	err := c.Run([]string{"getenvoy", "versions"})
+	t.Run("empty on missing versions dir", func(t *testing.T) {
+		rows, err := getInstalledVersions(homeDir)
+		require.NoError(t, err) // skips instead of crashing
+		require.Empty(t, rows)
+	})
 
-	require.NoError(t, err)
-	require.Equal(t, "No Envoy versions, yet\n", stdout.String())
-	require.Empty(t, stderr)
-}
+	versionsDir := filepath.Join(homeDir, "versions")
+	require.NoError(t, os.WriteFile(versionsDir, []byte{}, 0700))
 
-// TestGetEnvoyVersions verifies output is sorted
-func TestGetEnvoyVersions_Sorted(t *testing.T) {
-	o, cleanup := setupTest(t)
-	defer cleanup()
+	t.Run("error on file at versions dir", func(t *testing.T) {
+		_, err := getInstalledVersions(homeDir)
+		require.Error(t, err)
+	})
 
-	oneOneTwo := filepath.Join(o.HomeDir, "versions", "1.1.2")
+	require.NoError(t, os.Remove(versionsDir))
+	require.NoError(t, os.Mkdir(versionsDir, 0700))
+
+	t.Run("empty on empty versions dir", func(t *testing.T) {
+		rows, err := getInstalledVersions(homeDir)
+		require.NoError(t, err) // skips instead of crashing
+		require.Empty(t, rows)
+	})
+
+	oneOneTwo := filepath.Join(versionsDir, "1.1.2")
 	require.NoError(t, os.MkdirAll(oneOneTwo, 0700))
-	requireSetMtime(t, oneOneTwo, "2020-12-31")
+	morerequire.RequireSetMtime(t, oneOneTwo, "2020-12-31")
 
-	oneTwoOne := filepath.Join(o.HomeDir, "versions", "1.2.1")
-	require.NoError(t, os.MkdirAll(oneTwoOne, 0700))
-	requireSetMtime(t, oneTwoOne, "2020-12-30")
+	t.Run("release date from mtime", func(t *testing.T) {
+		rows, err := getInstalledVersions(homeDir)
+		require.NoError(t, err) // skips instead of crashing
+		require.EqualValues(t, rows, []versionReleaseDate{{"1.1.2", "2020-12-31"}})
+	})
 
-	oneTwoTwo := filepath.Join(o.HomeDir, "versions", "1.2.2")
-	require.NoError(t, os.MkdirAll(oneTwoTwo, 0700))
-	requireSetMtime(t, oneTwoTwo, "2020-12-31")
-
-	c, stdout, stderr := newApp(o)
-	o.Out = stdout
-	err := c.Run([]string{"getenvoy", "versions"})
-
-	require.NoError(t, err)
-	require.Equal(t, `VERSION	RELEASE_DATE
-1.2.2	2020-12-31
-1.1.2	2020-12-31
-1.2.1	2020-12-30
-`, stdout.String())
-	require.Empty(t, stderr)
-}
-
-func TestGetEnvoyVersions_DirIsFile(t *testing.T) {
-	o, cleanup := setupTest(t)
-	defer cleanup()
-
-	require.NoError(t, os.WriteFile(filepath.Join(o.HomeDir, "versions"), []byte{}, 0700))
-
-	c, stdout, stderr := newApp(o)
-	o.Out = stdout
-	err := c.Run([]string{"getenvoy", "versions"})
-
-	require.Error(t, err)
-	require.Empty(t, stdout)
-	require.Empty(t, stderr)
-}
-
-func TestGetEnvoyVersions_SkipsFiles(t *testing.T) {
-	o, cleanup := setupTest(t)
-	defer cleanup()
-
-	oneOneTwo := filepath.Join(o.HomeDir, "versions", "1.1.2")
-	require.NoError(t, os.MkdirAll(oneOneTwo, 0700))
-	requireSetMtime(t, oneOneTwo, "2020-12-31")
-
-	oneTwoOne := filepath.Join(o.HomeDir, "versions", "1.2.1")
+	oneTwoOne := filepath.Join(versionsDir, "1.2.1")
 	require.NoError(t, os.WriteFile(oneTwoOne, []byte{}, 0700)) // notice a file not a directory!
-	requireSetMtime(t, oneTwoOne, "2020-12-30")
+	morerequire.RequireSetMtime(t, oneTwoOne, "2020-12-30")
 
-	oneTwoTwo := filepath.Join(o.HomeDir, "versions", "1.2.2")
-	require.NoError(t, os.MkdirAll(oneTwoTwo, 0700))
-	requireSetMtime(t, oneTwoTwo, "2020-12-31")
-
-	c, stdout, stderr := newApp(o)
-	o.Out = stdout
-	err := c.Run([]string{"getenvoy", "versions"})
-
-	require.NoError(t, err)
-	require.Equal(t, `VERSION	RELEASE_DATE
-1.2.2	2020-12-31
-1.1.2	2020-12-31
-`, stdout.String())
-	require.Empty(t, stderr)
+	t.Run("skips file where version should be", func(t *testing.T) {
+		rows, err := getInstalledVersions(homeDir)
+		require.NoError(t, err) // skips instead of crashing
+		require.EqualValues(t, rows, []versionReleaseDate{{"1.1.2", "2020-12-31"}})
+	})
 }
 
-func TestGetEnvoyVersions_Remote(t *testing.T) {
-	o, cleanup := setupTest(t)
-	defer cleanup()
+func TestAddVersions(t *testing.T) {
+	goodVersions := map[string]version.EnvoyVersion{
+		"1.14.7": {
+			ReleaseDate: "2021-04-15",
+			Tarballs: map[string]string{
+				"darwin/amd64": "https://getenvoy.io/versions/1.14.7/envoy-1.14.7-darwin-x86_64.tar.gz",
+				"linux/amd64":  "https://getenvoy.io/versions/1.14.7/envoy-1.14.7-linux-x86_64.tar.gz",
+			},
+		},
+		"1.17.3": {
+			ReleaseDate: "2021-05-11",
+			Tarballs: map[string]string{
+				"linux/amd64": "https://getenvoy.io/versions/1.17.3/envoy-1.17.3-linux-x86_64.tar.gz",
+			},
+		},
+		"1.18.3": {
+			ReleaseDate: "2021-05-11",
+			Tarballs: map[string]string{
+				"darwin/amd64": "https://getenvoy.io/versions/1.18.3/envoy-1.18.3-darwin-x86_64.tar.gz",
+				"linux/amd64":  "https://getenvoy.io/versions/1.18.3/envoy-1.18.3-linux-x86_64.tar.gz",
+			},
+		},
+	}
 
-	// Run "getenvoy versions"
-	c, stdout, stderr := newApp(o)
-	o.Out = stdout
-	err := c.Run([]string{"getenvoy", "versions", "-a"})
+	tests := []struct {
+		name     string
+		existing []versionReleaseDate
+		update   map[string]version.EnvoyVersion
+		platform string
+		expected []versionReleaseDate
+	}{
+		{
+			name:     "darwin",
+			existing: []versionReleaseDate{},
+			update:   goodVersions,
+			platform: "darwin/amd64",
+			expected: []versionReleaseDate{{"1.14.7", "2021-04-15"}, {"1.18.3", "2021-05-11"}},
+		},
+		{
+			name:     "linux",
+			platform: "linux/amd64",
+			existing: []versionReleaseDate{},
+			update:   goodVersions,
+			expected: []versionReleaseDate{{"1.14.7", "2021-04-15"}, {"1.17.3", "2021-05-11"}, {"1.18.3", "2021-05-11"}},
+		},
+		{
+			name:     "already exists",
+			existing: []versionReleaseDate{{"1.14.7", "2020-01-01"}},
+			update:   goodVersions,
+			platform: "darwin/amd64",
+			expected: []versionReleaseDate{{"1.14.7", "2020-01-01"}, {"1.18.3", "2021-05-11"}},
+		},
+		{
+			name:     "unsupported OS",
+			existing: []versionReleaseDate{},
+			update:   goodVersions,
+			platform: "windows/amd64",
+			expected: []versionReleaseDate{},
+		},
+		{
+			name:     "unsupported Arch",
+			existing: []versionReleaseDate{},
+			update:   goodVersions,
+			platform: "linux/arm64",
+			expected: []versionReleaseDate{},
+		},
+		{
+			name:     "empty version list",
+			existing: []versionReleaseDate{},
+			platform: "darwin/amd64",
+			expected: []versionReleaseDate{},
+		},
+	}
 
-	require.NoError(t, err)
-	require.Equal(t, fmt.Sprintf(`VERSION	RELEASE_DATE
-%s	2020-12-31
-`, version.LastKnownEnvoy), stdout.String())
-	require.Empty(t, stderr)
+	for _, tt := range tests {
+		tc := tt
+		t.Run(tc.name, func(t *testing.T) {
+			require.NoError(t, addAvailableVersions(&tc.existing, tc.update, tc.platform))
+			require.ElementsMatch(t, tc.expected, tc.existing)
+		})
+	}
 }
 
-func TestGetEnvoyVersions_All_Mixed(t *testing.T) {
-	o, cleanup := setupTest(t)
-	defer cleanup()
+func TestAddVersions_Validates(t *testing.T) {
+	tests := []struct {
+		name   string
+		update map[string]version.EnvoyVersion
+	}{
+		{
+			name: "invalid releaseDate",
+			update: map[string]version.EnvoyVersion{
+				"1.14.7": {
+					ReleaseDate: "ice cream",
+					Tarballs: map[string]string{
+						"darwin/amd64": "https://getenvoy.io/versions/1.14.7/envoy-1.14.7-darwin-x86_64.tar.gz",
+					},
+				},
+			},
+		},
+		{
+			name: "missing releaseDate",
+			update: map[string]version.EnvoyVersion{
+				"1.14.7": {
+					Tarballs: map[string]string{
+						"darwin/amd64": "https://getenvoy.io/versions/1.14.7/envoy-1.14.7-darwin-x86_64.tar.gz",
+					},
+				},
+			},
+		},
+	}
 
-	oneOneTwo := filepath.Join(o.HomeDir, "versions", "1.1.2")
-	require.NoError(t, os.MkdirAll(oneOneTwo, 0700))
-	requireSetMtime(t, oneOneTwo, "2020-12-31")
-
-	// Run "getenvoy versions"
-	c, stdout, stderr := newApp(o)
-	o.Out = stdout
-	err := c.Run([]string{"getenvoy", "versions", "-a"})
-
-	require.NoError(t, err)
-	require.Equal(t, fmt.Sprintf(`VERSION	RELEASE_DATE
-%s	2020-12-31
-1.1.2	2020-12-31
-`, version.LastKnownEnvoy), stdout.String())
-	require.Empty(t, stderr)
-}
-
-func TestGetEnvoyVersions_All_PrefersInstallDate(t *testing.T) {
-	o, cleanup := setupTest(t)
-	defer cleanup()
-
-	installed := filepath.Join(o.HomeDir, "versions", version.LastKnownEnvoy)
-	require.NoError(t, os.MkdirAll(installed, 0700))
-	requireSetMtime(t, installed, "2021-12-31") // not 2020-12-31
-
-	// Run "getenvoy versions"
-	c, stdout, stderr := newApp(o)
-	o.Out = stdout
-	err := c.Run([]string{"getenvoy", "versions", "-a"})
-
-	require.NoError(t, err)
-	require.Equal(t, fmt.Sprintf(`VERSION	RELEASE_DATE
-%s	2021-12-31
-`, version.LastKnownEnvoy), stdout.String())
-	require.Empty(t, stderr)
-}
-
-func requireSetMtime(t *testing.T, dir, date string) {
-	td, err := time.Parse("2006-01-02", date)
-	require.NoError(t, err)
-	require.NoError(t, os.Chtimes(dir, td, td))
+	for _, tt := range tests {
+		tc := tt
+		t.Run(tc.name, func(t *testing.T) {
+			err := addAvailableVersions(&[]versionReleaseDate{}, tc.update, "darwin/amd64")
+			require.Error(t, err)
+			require.Contains(t, err.Error(), `invalid releaseDate of version "1.14.7" for platform "darwin/amd64":`)
+		})
+	}
 }
